Rename mark_identity to alias_analysis.
